Selva Oscura: clang-tidy fixes (mechanical) + 4 of 7 cognitive-complexity offenders refactored. Drop redundant = {} member inits in Weapon.h / WeaponClass.h (13 sites). Const-correct loop iterators in PoseSampler.cpp + Tunables.cpp + SkeletalMesh.cpp + PerFrameTick.cpp + SceneGeometry.cpp. emplace_back instead of push_back at 2 sites in PoseSampler. Range-based for over bone_weights[4] in SkeletalMesh. Narrowing/widening conversion casts in SkeletalAssets/PerFrameTick/SceneGeometry. Delete dead-store legacy phase-match metrics (prev_is_gait, new_is_gait, prev_dur, target init, loco_lockout). Cognitive-complexity refactors: clipJointMotionEnd / clipJointMotionStart / clipJointMotionPeak share collectJointXYZSamples + per-joint scan helpers (was CCN 36/32/21, all under threshold now). loadSkeletalMesh split into readMeshAttributes + computeAssetRoot + buildInverseBindMatrices + bakeVertex + uploadMeshToGpu (was CCN 47, now under threshold). Remaining 3 cognitive offenders (PoseSampler::update, selvaPerFrame, resolveAttackCancelOpenTimes) are state-machine code coupled to the leg-spasm fixes; will refactor in separate commits each so regressions can be bisected.

+++ games/selva-oscura/src/anim/SkeletalAssets.cpp
@@ -151,7 +151,7 @@ void auditClipHipMotion()
         float cumul = 0.0f;
         for (int i = 0; i <= 10; ++i)
         {
-            const float t = (i / 10.0f) * dur;
+            const float t = (static_cast<float>(i) / 10.0f) * dur;
             const glm::vec2 hip = sSampler.sampleHipXZAt(*clip, t);
             const float step = (i == 0) ? 0.0f : glm::length(hip - prev);
             cumul += step;
